coffee_shop/requirements.txt
asgiref==3.4.1
Cerberus==1.3.4
Django==3.2.12
djangorestframework==3.13.1
pkg-resources==0.0.0
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
